// frontend/src/types/generated/financial/models/payableAccountCreate.ts
export interface PayableAccountCreate {
  /**
   * @minLength 3
   * @maxLength 500
   */
  description: string;
  document_number?: PayableAccountCreateDocumentNumber;
  payable_type?: PayableType;
  priority?: PayablePriority;
  supplier_id?: PayableAccountCreateSupplierId;
  category_id?: PayableAccountCreateCategoryId;
  gross_value: PayableAccountCreateGrossValue;
  discount_value?: PayableAccountCreateDiscountValue;
  addition_value?: PayableAccountCreateAdditionValue;
  withhold_iss?: PayableAccountCreateWithholdIss;
  withhold_ir?: PayableAccountCreateWithholdIr;
  withhold_pis?: PayableAccountCreateWithholdPis;
  withhold_cofins?: PayableAccountCreateWithholdCofins;
  withhold_csll?: PayableAccountCreateWithholdCsll;
  withhold_inss?: PayableAccountCreateWithholdInss;
  issue_date?: PayableAccountCreateIssueDate;
  due_date: string;
  competence_date?: PayableAccountCreateCompetenceDate;
  /**
   * @minimum 1
   * @maximum 360
   */
  total_installments?: number;
  is_recurring?: boolean;
  recurrence_type?: PayableAccountCreateRecurrenceType;
  recurrence_end_date?: PayableAccountCreateRecurrenceEndDate;
  cost_center?: PayableAccountCreateCostCenter;
  project?: PayableAccountCreateProject;
  fiscal_document_type?: PayableAccountCreateFiscalDocumentType;
  fiscal_document_key?: PayableAccountCreateFiscalDocumentKey;
  fiscal_document_url?: PayableAccountCreateFiscalDocumentUrl;
  contract_id?: PayableAccountCreateContractId;
  payment_method_id?: PayableAccountCreatePaymentMethodId;
  bank_account_id?: PayableAccountCreateBankAccountId;
  requires_approval?: boolean;
  tags?: string[];
  notes?: PayableAccountCreateNotes;
  condominio_id: string;
}